mender-custom-flash-layout: fix jetson-nano-qspi-sd layout
which is now using a reduced size for the U-Boot environment
area, leaving the VER_b and VER partitions in their L4T default
locations in the SPI flash.

This is a change in L4T R32.4.3 that has not yet been
accounted for upstream in the dunfell branch of
meta-mender-community/meta-mender-tegra.

Signed-off-by: Matt Madison <matt@madison.systems>